+++ backend/internal/handlers/logout.go
@@ -1,8 +1,6 @@
 package handlers
 
 import (
-	"fmt"
-
 	"github.com/alireza/api/internal/database"
 	sessionServices "github.com/alireza/api/internal/services/session"
 	"github.com/gin-gonic/gin"
@@ -10,8 +8,7 @@ import (
 
 func Logout(c *gin.Context) {
 	s := sessionServices.New()
-	session, _ := c.Get("session")
-	fmt.Println(session)
+	session := c.GetString("session")
 	err := s.DeleteSession(database.DB, session)
 	if err != nil {
 		c.JSON(500, err)