+++ backend/internal/handlers/register.go
@@ -1,17 +1,12 @@
 package handlers
 
 import (
-	"fmt"
-	"log"
-	"time"
-	"unicode"
-
 	"github.com/alireza/api/internal/database"
 	"github.com/alireza/api/internal/models"
-	uServices "github.com/alireza/api/internal/services/user"
+	userService "github.com/alireza/api/internal/services/user"
+	"github.com/alireza/api/internal/utils"
+	"github.com/alireza/api/internal/validators"
 	"github.com/gin-gonic/gin"
-	"github.com/golodash/galidator"
-	"golang.org/x/crypto/bcrypt"
 )
 
 type RegisterRequest struct {
@@ -21,70 +16,38 @@ type RegisterRequest struct {
 }
 
 func Register(c *gin.Context) {
-	g := galidator.New()
-	validator := g.Validator(galidator.Rules{
-		"Username": g.RuleSet().Required().Min(5).Max(64),
-		"Password": g.RuleSet().Required().Min(8).Max(100),
-		"Email":    g.RuleSet().Required().Max(320),
-	})
 	req := &RegisterRequest{}
-	if err := c.BindJSON(&req); err != nil {
-		c.JSON(400, gin.H{
-			"message": "Fill blank fields",
-		})
+	s := userService.New()
+
+	// Parse json
+	if !utils.ParseJson(req, c) {
 		return
 	}
-	err := validator.Validate(req)
-	fmt.Println(err)
-	if err != nil {
-		c.JSON(400, gin.H{"errors : ": err})
+
+	// Validate user input
+	if err := validators.RegisterValidator.Validate(*req); err != nil {
+		c.JSON(400, gin.H{"errors": err})
 		return
 	}
-	// Registering validations
-	// if len(req.Username) < 5 || len(req.Username) > 64 {
-	// 	c.JSON(400, gin.H{
-	// 		"message": "Username should be 5-64 characters long",
-	// 	})
-	// 	return
-	// }
-	for _, char := range req.Username {
-		if !unicode.IsLetter(char) && !unicode.IsNumber(char) {
-			c.JSON(400, gin.H{
-				"message": "Only alphanumeric characters are allowed",
-			})
-			return
-		}
-	}
-	// if len(req.Password) < 8 || len(req.Password) > 100 {
-	// 	c.JSON(400, gin.H{
-	// 		"message": "Password should be 8-100 characters long",
-	// 	})
-	// 	return
-	// }
-	// Importing user services
-	s := uServices.New()
-	if s.UserExists(database.DB, req.Username, req.Email) {
-		c.JSON(400, gin.H{
-			"message": "Username or Email already taken, try another",
-		})
-		return
-	}
+
 	// Hashing password
-	hash, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcrypt.DefaultCost)
+	hashedPassword, err := s.HashPassword(req.Password)
 	if err != nil {
-		log.Fatal(err)
+		c.JSON(500, gin.H{"errors": err})
+		return
 	}
+
 	// Creating user
 	model := &models.User{
-		Username:  req.Username,
-		Password:  string(hash),
-		Email:     req.Email,
-		CreatedAt: time.Now().Local(),
+		Username: req.Username,
+		Password: hashedPassword,
+		Email:    req.Email,
 	}
 	user, err := s.CreateUser(database.DB, *model)
 	if err != nil {
 		c.JSON(500, err)
 		return
 	}
-	c.JSON(200, user)
+
+	c.JSON(201, user)
 }