+++ backend/internal/validators/init.go
@@ -1,0 +1,5 @@
+package validators
+
+import "github.com/golodash/galidator"
+
+var g = galidator.New()